chumby8: add SD card support
This was mostly just populating the device tree based on the memory map.
Also, the mv_sdhci driver was only available on the MVEBU arch, so I had
to allow it on Chumby as well.

According to all of my tests, this controller is affected by the "broken
R1B" quirk. I found that it was always timing out on R1b responses.
After adding this quirk, it seems to behave fine, other than the
annoying 1 second timeout after commands.