sdhci: ignore INT_DATA_END with no data if broken R1b quirk is present
This has gone back and forth a couple of times. The existing behavior
causes issues with the mv_sdhci driver. I have come to the "happy
medium" solution of only ignoring INT_DATA_END if SDHCI_QUIRK_BROKEN_R1B
is present. This fixes the issue on mv_sdhci, and hopefully limits
interference with other controllers.