chumby8: port PLL and DDR reconfiguration routines from old u-boot
As part of this process, I cleaned up the code and removed a bunch of
debug output so that the u-boot boot process appears nice and clean.

It's unclear how much of these algorithms are actually needed; there is
a lot of register reading going on, but the original u-boot code said it
was working around silicon bugs. So I left it all in there.