chumby8: add PWM2 initialization
I ran into a lot of problems with the clock/reset bits for PWM2. PWM1
and PWM2 share common clock/reset bits in the PWM1 register, and there
is some undocumented magic involved in powering it on properly so that
PWM2 works when PWM1 isn't turned on. The shared clock doesn't work
unless you kickstart it by turning on PWM1 briefly.