chumby8: add quirk to disable card reader by default
The card reader shows up as a disk, even if no card is inserted. This
means that you always get an "sda" device. Prevent the usb-storage
driver from binding to it by adding a quirk.

A userspace utility can remove the quirk and monitor the card detect
pins to decide when to attach and remove the card reader device.